Arreglando erroresv2 :)
