+++ src/main/java/org/springframework/samples/nt4h/card/ability/rogueEffects/SaqueoOroGloria.java
@@ -1,4 +1,4 @@
-package org.springframework.samples.nt4h.card.ability.rogue;
+package org.springframework.samples.nt4h.card.ability.rogueEffects;
 
 import org.springframework.samples.nt4h.card.enemy.EnemyInGame;
 import org.springframework.samples.nt4h.game.Game;
@@ -8,11 +8,12 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public class SaqueoOro {
+public class SaqueoOroGloria {
     public void execute(Player player) {
         Game game = player.getGame();
-        List<EnemyInGame> activeOrcs = game.getOrcs();
+        List<EnemyInGame> activeOrcs = game.getActualOrcs();
         Integer enemies = activeOrcs.size();
-        new GiveGold(enemies * 2, player).execute();
+        new StealCoin(enemies, player).execute();
+        new GainGlory(player).execute();
     }
 }
