--- src/main/java/org/springframework/samples/nt4h/card/ability/warrior/CargaConEscudo.java
@@ -1,13 +1,0 @@
-package org.springframework.samples.nt4h.card.ability.warrior;
-
-import org.springframework.samples.nt4h.card.enemy.EnemyInGame;
-import org.springframework.samples.nt4h.player.Player;
-import org.springframework.stereotype.Component;
-
-@Component
-public class CargaConEscudo {
-    public void execute(Player player, EnemyInGame enemy) {
-        new Attack(2, player, enemy).execute();
-        new PreventDamage(2, player).execute();
-    }
-}
