+++ src/main/java/org/springframework/samples/nt4h/card/ability/warriorEffects/AtaqueBrutal.java
@@ -1,14 +1,13 @@
-package org.springframework.samples.nt4h.card.ability.wizard;
+package org.springframework.samples.nt4h.card.ability.warriorEffects;
 
 import org.springframework.samples.nt4h.card.enemy.EnemyInGame;
 import org.springframework.samples.nt4h.player.Player;
 import org.springframework.stereotype.Component;
 
 @Component
-public class DisparoGelido {
+public class AtaqueBrutal {
     public void execute(Player player, EnemyInGame enemy) {
-        new Attack(1, player, enemy).execute();
-        new DrawCards(1, player).execute();
-        new DenyDamageEnemy(enemy).execute();
+        new Attack(3, player, enemy).execute();
+        new DiscardCards(1, player).execute();
     }
 }
