+++ src/main/java/org/springframework/samples/nt4h/card/ability/warriorEffects/DobleEspadazo.java
@@ -1,13 +1,13 @@
-package org.springframework.samples.nt4h.card.ability.wizard;
+package org.springframework.samples.nt4h.card.ability.warriorEffects;
 
 import org.springframework.samples.nt4h.card.enemy.EnemyInGame;
 import org.springframework.samples.nt4h.player.Player;
 import org.springframework.stereotype.Component;
 
 @Component
-public class ProyectilIgneo {
+public class DobleEspadazo {
     public void execute(Player player, EnemyInGame enemy) {
         new Attack(2, player, enemy).execute();
-        new GainGlory(1, player).execute();
+        new DiscardCards(1, player).execute();
     }
 }
