+++ src/main/java/org/springframework/samples/nt4h/card/ability/warriorEffects/VozDeAliento.java
@@ -1,4 +1,4 @@
-package org.springframework.samples.nt4h.card.ability.warrior;
+package org.springframework.samples.nt4h.card.ability.warriorEffects;
 
 import org.springframework.samples.nt4h.game.Game;
 import org.springframework.samples.nt4h.player.Player;
