+++ src/main/java/org/springframework/samples/nt4h/card/ability/wizardEffects/FlechaCorrosiva.java
@@ -1,4 +1,4 @@
-package org.springframework.samples.nt4h.card.ability.warrior;
+package org.springframework.samples.nt4h.card.ability.wizardEffects;
 
 import org.springframework.samples.nt4h.card.ability.AbilityEffectEnum;
 import org.springframework.samples.nt4h.card.enemy.EnemyInGame;
@@ -6,16 +6,10 @@ import org.springframework.samples.nt4h.player.Player;
 import org.springframework.stereotype.Component;
 
 @Component
-public class Espadazo {
+public class FlechaCorrosiva {
     public void execute(Player player, EnemyInGame enemy) {
         new Attack(1, player, enemy).execute();
-
-        if (!enemy.getPermanentEffectCardsUsed().contains(AbilityEffectEnum.ESPADAZO)) {
-            new DrawCards(1, player).execute();
-        }
-
-        player.getGame().getOrcs()
-            .forEach(x -> x.getPermanentEffectCardsUsed().add(AbilityEffectEnum.ESPADAZO));
-
+        new DiscardCards(1, player).execute();
+        enemy.getPermanentEffectCardsUsed().add(AbilityEffectEnum.FLECHA_CORROSIVA);
     }
 }
