+++ src/main/java/org/springframework/samples/nt4h/card/ability/wizardEffects/Reconstitucion.java
@@ -1,4 +1,4 @@
-package org.springframework.samples.nt4h.card.ability.wizard;
+package org.springframework.samples.nt4h.card.ability.wizardEffects;
 
 import org.springframework.samples.nt4h.player.Player;
 import org.springframework.stereotype.Component;
